query_withdrawable_unbonded takes into account slashing witch may had happened #LSC-299
